+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added initial support for I/O (sockets and files).
+- Re-added support for `$` to access the function parameter.
+
 ## [0.3.0] - 2025-10-31
 
 ### Added